+++ src/app/@application/page.js
@@ -1,5 +1,7 @@
 import CompleteProfile from "./(hoc)/completeProfile";
-
+export const metadata = {
+    title: 'Keepingly'
+}
 export default function homePage (){
     return(
         <CompleteProfile />
